pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fuzzy_cnn"
version = "0.1.0"
description = "A tech test project for building a CIFAR-10 image classifier"
requires-python = ">=3.9"
authors = [
    { name="David A'Hearne", email="david.ahearnex@gmail.com" }
]

dependencies = [
    "pydantic",
    "pydantic-settings",
    "pillow"
]

[project.optional-dependencies]
serve = [
    "fastapi[standard]",
    "onnx",
    "onnxruntime",
    "onnxscript"
]

train = [
    "torch",
    "torchvision"
]

test = [
    "pytest"
]

[project.scripts]
fuzzy_serve = "fuzzy_cnn.serve.api:main"
fuzzy_train = "fuzzy_cnn.train.train:train_model"
fuzzy_export = "fuzzy_cnn.export.export_onnx:export_onnx"